+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.1] - 2023-08-11
 ### Fixed
-- This fixes a small issue with the regex used to extract the time for e.g. break tags.
+- Regex parsing of decibels and unsigned percentages
+- Parsing of empty pitch contours
+- Serialization of Break strength
+
+### Changed
+- Moved sign type from character to enum
+
+## [0.1.0] - 2023-07-19 
 ### Added 
 - Added attributes for say-as, prosody and emphasis tags and functions for parsing those
 - Added basic parsing of elements with no extracting of attributes for standard elements
@@ -18,3 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reject invalid nesting of elements and add API functions too check if elements can be nested
 - Description element text is now ignored
 - Ability to expand sub elements during SSML parsing
+
+### Fixed
+- This fixes a small issue with the regex used to extract the time for e.g. break tags.